% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ateITS.r
\name{ateITS}
\alias{ateITS}
\title{ateITS
Computes average treatment effect of the policy that can be aggregated in the desired way.}
\usage{
ateITS(forecast.object, ite.object, n.periods = 1, groups = NULL)
}
\arguments{
\item{forecast.object}{object  from \code{forecastITS}}

\item{ite.object}{object from \code{iteITS}}

\item{n.periods}{number of periods post treatment for TATE. Default = 1}

\item{groups}{a vector for which groups are investigated. Default is NULL: ate for the
population are computed}
}
\description{
GATE and TATE and InstATE
Group average treatment effect and time average treatment effect and instantaneous
treatment effect
}
