% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidateITS.R
\name{crossValidateITS}
\alias{crossValidateITS}
\title{Cross validation function}
\usage{
crossValidateITS(data, id, k)
}
\arguments{
\item{data}{Data. Will be converted into a ´data.frame'}

\item{id}{Unique identifier}

\item{k}{Number of cross-validation folds}
}
\description{
Cross validation function
}
\examples{
df <- data.frame(ind = sample(1:20, 200, replace = TRUE),
                 y = rnorm(200, mean = 10, sd = 5))
crossValidateITS(df, id = "ind", k = 5)

}
