% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandITS.R
\name{expandITS}
\alias{expandITS}
\title{Create a balanced panel if many 0}
\usage{
expandITS(data, key, y, time, covariates_time = NULL, covariates_fix = NULL)
}
\arguments{
\item{data}{data is in long format (multiple rows per key)}

\item{key}{is the key that identifies units of observations.}

\item{y}{is a string indicating the series.}

\item{time}{is a string indicating the time variable in data.}

\item{covariates_time}{is a vector of variable names including the covariates that
change with time. For better prediction, it is important to include in this vector
the time variables (for instance, "year" or "season")}

\item{covariates_fix}{is a vector of individual variables that are time invariant.
the \code{forecast} does not yet implement time-varying covariates (i.e. series of other
variables).}
}
\description{
Wrapper around \code{tidyr::expand} and \code{left_join} to create balanced panels.
}
\examples{

time <- rep(c(1:50), 5)
id   <- rep(1:5, each = 50)
year <- rep(c(rep(1990, 12), rep(1991, 12), rep(1992, 12), rep(1993, 12), rep(1994, 2)), 5)
y    <- time + rnorm(50, 0, 1) + (time>40)*rnorm(10, 2, 2)
df   <- data.frame(id, time, year, y)
df   <- df[sample(1:250, 230), ]
ex   <- expandITS(data = df, time = "time", covariates_time = c("year"), key = "id", 
                y = "y")
}
