% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factordummyITS.r
\name{factor_to_dummy}
\alias{factor_to_dummy}
\title{factor_to_dummy}
\usage{
factor_to_dummy(data = data.frame(), factor_vector_names = c())
}
\arguments{
\item{data}{a data.frame or a data.table}

\item{factor_vector_names}{a vector of strings for the names of variables
needed to be converted.}
}
\value{
a data frame
}
\description{
Function to transform dummy variables into factors.
}
\examples{
data <- data.frame("y" = sample(c(0, 0:2), 1000, TRUE),
                   "year" = c(rep(1:3, 333), 3),
                   "age" = sample(c(18, 25, 30, 40, 50, 60), 1000, TRUE),
                   "gender" = sample(c(0,1), 1000, TRUE),
                   "region" = sample(1:5, 1000, TRUE)
                   )

}
