% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Multi-step forecasting function}
\usage{
fit(
  x.train,
  y.train,
  x.test,
  y.test,
  binary = F,
  method = c("lasso", "lm", "rf", "xgboost")
)
}
\arguments{
\item{x.train}{the matrix of predictors from the train dataset}

\item{y.train}{the vector of target variable from the train dataset}

\item{x.test}{the matrix of predictors from the test dataset}

\item{y.test}{the vector of target variable from the test dataset}

\item{binary}{binary variable indicating whether the target \code{y} is a binary
variable. If yes, ...}

\item{method}{vector of method names. Methods available are "lm", "rf", "lasso".}
}
\description{
Multi-step forecasting function
}
\examples{
n <- 1000
x <- matrix(rnorm(n * 4), ncol = 4)
colnames(x) <- paste0("x", 1:4)

y <- x[, 1] + 2 * x[, 2] + rnorm(n)

x.train <- x[1:750,]
y.train <- y[1:750]
x.test  <- x[751:1000,]
y.test  <- y[751:1000]

fit(x.train, y.train, x.test, y.test, method = "lm")
}
\keyword{internal}
