% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitOLD}
\alias{fitOLD}
\title{Multi-step forecasting function}
\usage{
fitOLD(
  x.train,
  y.train,
  x.test,
  y.test,
  binary = F,
  method = c("lasso", "lm", "rf")
)
}
\arguments{
\item{x.train}{the matrix of predictors from the train dataset}

\item{y.train}{the vector of target variable from the train dataset}

\item{x.test}{the matrix of predictors from the test dataset}

\item{y.test}{the vector of target variable from the test dataset}

\item{binary}{binary variable indicating whether the target \code{y} is a binary
variable. If yes, ...}

\item{method}{vector of method names. Methods available are "lm", "rf", "lasso".}
}
\description{
Multi-step forecasting function
}
\keyword{internal}
