% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenDataITS.R
\name{flattenDataITS}
\alias{flattenDataITS}
\title{flattenDataITS}
\usage{
flattenDataITS(
  data,
  index,
  WINDOW,
  STEPS,
  time,
  covariates_time,
  covariates_fix,
  key,
  outcome
)
}
\arguments{
\item{outcome}{}
}
\description{
flattenDataITS
}
