% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastITS.R
\name{forecastITS}
\alias{forecastITS}
\title{Forecast function for ITS}
\usage{
forecastITS(
  data,
  time,
  INDEX = 0L,
  WINDOW = 12L,
  STEPS = as.integer(WINDOW/3),
  covariates_time,
  covariates_fix = NULL,
  key,
  y,
  method = c("lm", "rf"),
  K = 5,
  CYCLE = 12L,
  FORECASTUNITS = NULL
)
}
\arguments{
\item{data}{data is in long format (multiple rows per key)}

\item{time}{is a string indicating the time variable in data.}

\item{INDEX}{is an integer indicating the time point where the treatment begins.
Default is index = 0, with pre-treatment periods having time < 0.}

\item{WINDOW}{is an integer indicating the number of time periods before time t
are used for prediction. Ideally, if time is in month, WINDOW should include a full
seasonality cycle (i.e. WINDOW = 12 if in months)}

\item{STEPS}{is an integer indicating the number of time periods used to assess fit
before the intervention.}

\item{covariates_time}{is a vector of variable names including the covariates that
change with time. For better prediction, it is important to include in this vector
the time variables (for instance, "year" or "season")}

\item{covariates_fix}{is a vector of individual variables that are time invariant.
the \code{forecast} does not yet implement time-varying covariates (i.e. series of other
variables).}

\item{key}{is the key that identifies units of observations.}

\item{y}{is a string indicating the series.}

\item{method}{is a vector containing the prediction methods used for prediction.
Possible methods are c("lm", "rf").}

\item{K}{is the number of folds for the k-fold cross-validation technique.
Default is K = 5.}
}
\description{
Main function for the ITS. Uses cross-validation to perform "sliding-window" in
single-step or multiple-step estimation of time series.
}
