% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteITS.R
\name{iteITS}
\alias{iteITS}
\title{iteITS
Computes individual treatment effect of the policy that can be aggregated in the desired way.}
\usage{
iteITS(forecast.object)
}
\arguments{
\item{forecast.object}{object  from \code{forecastITS}}
}
\description{
iteITS
Computes individual treatment effect of the policy that can be aggregated in the desired way.
}
