% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itscausal-package.R
\docType{package}
\name{itscausal-package}
\alias{itscausal}
\alias{itscausal-package}
\title{itscausal: Interrupted Time Series for Causal Estimation}
\description{
This package performs single- and multiple-steps forecasting for estimating causal effects using time series. It relies on machine learning methods for prediction with crossfitting and ensemble learning.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ASallin/itscausal}
  \item Report bugs at \url{https://github.com/ASallin/itscausal/issues}
}

}
\author{
\strong{Maintainer}: Aurelien Sallin \email{aurelien.sallin@protonmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
