% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED_lm_fit.r
\name{lm_fit}
\alias{lm_fit}
\title{This function estimates linear regressions based on the \code{\link{lm}} function}
\usage{
lm_fit(x, y, args = list())
}
\arguments{
\item{x}{Matrix of covariates (number of observations times number of covariates matrix)}

\item{y}{vector of outcomes (no df format)}

\item{args}{List of arguments passed to  \code{\link{ranger}}}
}
\description{
This function estimates linear regressions based on the \code{\link{lm}} function
}
\keyword{internal}
