% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_class.r
\docType{class}
\name{MLModelITS-class}
\alias{MLModelITS-class}
\title{MLModelITS Class}
\description{
This class contains a ML model for prediction. It allows to train a model,
and to predict on a test sample.
}
\section{Slots}{

\describe{
\item{\code{model}}{is the model contained in the class.}

\item{\code{type}}{gives the type of the ML algorithm used.}
}}

