% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_class.r
\name{MLModelITS}
\alias{MLModelITS}
\title{Constructor function for creating MLModel objects}
\usage{
MLModelITS(model, type)
}
\arguments{
\item{model}{is the model}

\item{type}{is the type of algorithm used}
}
\description{
Creates instances of class MLModelITS
}
