% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fit.r
\name{predict.lm_fit}
\alias{predict.lm_fit}
\title{Prediction based on ranger}
\usage{
\method{predict}{lm_fit}(lm_fit, x, xnew = NULL)
}
\arguments{
\item{lm_fit}{Output of \code{\link{ranger}} or \code{\link{ranger_fit}}}

\item{x}{Covariate matrix of training sample}

\item{xnew}{Covariate matrix of test sample}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for xnew}
}
\description{
Prediction based on ranger
}
\keyword{internal}
