% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_class.r
\name{predict.MLModelITS}
\alias{predict.MLModelITS}
\title{predict.MLModelITS}
\usage{
\method{predict}{MLModelITS}(object, x, xnew = NULL, ...)
}
\arguments{
\item{object}{Output of \code{\link{ranger}} or \code{\link{ranger_fit}}}

\item{x}{Covariate matrix of training sample}

\item{xnew}{Covariate matrix of test sample}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for xnew}
}
\description{
predict.MLModelITS
}
\keyword{internal}
