% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED_ranger_fit.r
\name{predict.ranger_fit}
\alias{predict.ranger_fit}
\title{Prediction based on ranger}
\usage{
\method{predict}{ranger_fit}(ranger_fit, x, xnew = NULL)
}
\arguments{
\item{x}{Covariate matrix of training sample}

\item{xnew}{Covariate matrix of test sample}

\item{lasso_fit}{Output of \code{\link{ranger}} or \code{\link{ranger_fit}}}
}
\value{
Returns list containing:
\item{prediction}{vector of predictions for xnew}
}
\description{
Prediction based on ranger
}
\keyword{internal}
