% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED_stepcastITS.R
\name{stepcastITS}
\alias{stepcastITS}
\title{stepcastITS}
\usage{
stepcastITS(
  models,
  x,
  STEPS = 1L,
  RMSEweights,
  covariates_fix,
  covariates_time
)
}
\arguments{
\item{RMSEweights}{}
}
\description{
stepcastITS
}
\keyword{internal}
