% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_class.r
\name{train.MLModelITS}
\alias{train.MLModelITS}
\title{train.MLModelITS}
\usage{
\method{train}{MLModelITS}(object, x, y, tuning_params = NULL, ...)
}
\arguments{
\item{object}{takes object of class \code{MLModelITS}}

\item{x}{a matrix of predictors}

\item{y}{a vector of target}

\item{tuning_params}{a list containing tuning parameters for algorithms}
}
\description{
Define the train function as an S3 generic
}
