% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleLearnerITS.r
\name{ensemble}
\alias{ensemble}
\title{Ensemble learner function
computes a grid search of all possible weights that minimizes the RMSE within
each K folds. See Chernozhukov et al. for implementation.}
\usage{
ensemble(predictors, k, y = NULL, RMSEweights = NULL)
}
\arguments{
\item{predictors}{is a matrix of predictors}

\item{k}{is the number of predictors that should be included in the learning
computation.}

\item{y}{is the observed y. Set to NULL if a vector of weights is given.}
}
\description{
Ensemble learner function
computes a grid search of all possible weights that minimizes the RMSE within
each K folds. See Chernozhukov et al. for implementation.
}
