% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenDataITS.R
\name{flattenDataITS}
\alias{flattenDataITS}
\title{flattenDataITS}
\usage{
flattenDataITS(
  data,
  index,
  WINDOW,
  STEPS,
  time,
  covariates_time,
  covariates_fix,
  covariates_var = NULL,
  key,
  y
)
}
\arguments{
\item{data}{Dataset}

\item{index}{Index for intervention}

\item{WINDOW}{Window for prediction}

\item{y}{}
}
\description{
flattenDataITS
}
