% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastITS.R
\name{forecastITS}
\alias{forecastITS}
\title{Forecast function for ITS}
\usage{
forecastITS(
  data,
  time,
  key,
  y,
  INDEX = 0L,
  WINDOW = NULL,
  STEPS = NULL,
  covariates_time,
  covariates_fix = NULL,
  covariates_var = NULL,
  method = c("lm", "rf", "xgboost"),
  K = 5,
  FORECASTUNITS = NULL
)
}
\arguments{
\item{data}{data is in long format (multiple rows per key)}

\item{time}{is a string indicating the time variable in data.}

\item{key}{is the key that identifies units of observations.}

\item{y}{is a string indicating the series.}

\item{INDEX}{is an integer indicating the time point where the treatment begins.
Default is index = 0, with pre-treatment periods having time < 0.}

\item{WINDOW}{is an integer indicating the number of time periods before time t
are used for prediction. Ideally, if time is in month, WINDOW should include a full
seasonality cycle (i.e. WINDOW = 12 if in months)}

\item{STEPS}{is an integer indicating the number of time periods used to assess fit
before the intervention.}

\item{covariates_time}{is a vector of variable names including the covariates that
change with time. For better prediction, it is important to include in this vector
the time variables (for instance, "year" or "season")}

\item{covariates_fix}{is a string vector of individual variables that are time invariant.}

\item{covariates_var}{is a string vector of individual variables that are time variant.}

\item{method}{is a vector containing the prediction methods used for prediction.
Possible methods are c("lm", "rf"m "xgboost").}

\item{K}{is the number of folds for the k-fold cross-validation technique.
Default is K = 5.}
}
\description{
Main function for the ITS. Uses cross-validation to perform "sliding-window" in
single-step or multiple-step estimation of time series.
}
\examples{

time <- rep(c(1:50), 5)
id <- rep(1:5, each = 50)
year <- rep(c(rep(1990, 12), rep(1991, 12), rep(1992, 12), rep(1993, 12), rep(1994, 2)), 5)
y <- time + rnorm(50, 0, 1) + (time > 40) * rnorm(10, 2, 2)
x1 <- rep(sample(c(0, 1), 5, replace = T), each = 50)
x2 <- rep(sample(c(1, 2, 3), 5, replace = T), each = 50)
xv1 <- rep(sample(c(0, 1), 50, replace = T), 5)
xv2 <- rep(sample(c(1, 2, 3), 50, replace = T), 5)
df <- cbind(time, id, year, y, x1, x2, xv1, xv2)

f <- forecastITS(
  data = df, time = "time", INDEX = 40, covariates_time = c("year"), key = "id",
  y = "y"
)

f1 <- forecastITS(
  data = df, time = "time", INDEX = 40, covariates_time = c("year"), key = "id",
  y = "y",
  covariates_fix = c("x1", "x2"),
  covariates_var = c("xv1", "xv2")
)

# Unbalanced panel
dfUB <- df[sample(1:250, 230), ]
fUB <- forecastITS(
  data = df, time = "time", INDEX = 40, covariates_time = c("year"), key = "id",
  y = "y"
)
}
